Save and restore AI query pane state on chat close/open
